backend/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.1.0
supabase>=2.0.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
httpx>=0.25.0
